+++ .gitignore
@@ -1,8 +1,9 @@
-bower_components/
 node_modules/
-src/bower_components/
 dist/
 build/
 _SpecRunner.html
 src/js/.tern-port
 .DS_Store
+volumes/
+.vscode/
+.devcontainer/
