+++ package.json
@@ -10,6 +10,7 @@
     "babel-loader": "^8.0.6",
     "babel-plugin-add-module-exports": "^1.0.2",
     "babel-plugin-istanbul": "^6.0.0",
+    "babel-plugin-syntax-dynamic-import": "^6.18.0",
     "babel-preset-stage-0": "^6.24.1",
     "babelify": "^10.0.0",
     "browserify": "^16.5.0",
@@ -17,7 +18,7 @@
     "browserify-istanbul": "^3.0.1",
     "chai": "^4.2.0",
     "chai-as-promised": "^7.1.1",
-    "grunt": "^1.0.4",
+    "grunt": "^1.4.0",
     "grunt-babel": "^8.0.0",
     "grunt-contrib-clean": "^2.0.0",
     "grunt-contrib-compress": "^1.6.0",
@@ -43,23 +44,38 @@
     "karma-junit-reporter": "^2.0.1",
     "karma-mocha": "^2.0.1",
     "karma-mocha-reporter": "^2.2.5",
+    "libxmljs2": "^0.26.7",
     "mocha": "^7.0.1",
+    "mock-applicationmashup": "*",
     "sinon": "^8.1.1",
     "watchify": "^3.11.1",
     "webpack": "^4.41.5",
+    "webpack-cli": "^4.6.0",
+    "webpack-dev-server": "^3.11.2",
     "wirecloud-config-parser": "^0.2.3"
   },
   "private": true,
   "dependencies": {
+    "@deck.gl/core": "^8.4.13",
+    "@deck.gl/geo-layers": "^8.4.13",
+    "@deck.gl/mapbox": "^8.4.13",
+    "@deck.gl/react": "^8.4.13",
+    "@loaders.gl/3d-tiles": "^2.3.13",
     "@turf/helpers": "^6.3.0",
     "@turf/turf": "^6.3.0",
     "css-loader": "^5.2.4",
+    "deckgl-gsi-terrain-layer": "^1.0.1",
+    "liftoff": "^2.5.0",
     "maplibre-gl": "^1.14.0",
+    "math.gl": "^3.4.2",
     "style-loader": "^2.0.0",
-    "three": "^0.116.0",
+    "three": "^0.116.1",
     "three-gltf-loader": "^1.111.0",
     "topojson-client": "^3.1.0",
     "topojson-server": "^3.0.1",
     "topojson-simplify": "^3.0.3"
+  },
+  "scripts": {
+    "debug": "webpack-cli serve"
   }
 }
