+++ script/Mapbox3DTiles.patch
@@ -1,0 +1,16 @@
+@@ -1,3 +1,6 @@
++import * as THREE from 'three';
++import { GLTFLoader } from 'three/examples/jsm/loaders/GLTFLoader.js';
++
+ const MERCATOR_A = 6378137.0;
+ const WORLD_SIZE = MERCATOR_A * Math.PI * 2;
+
+@@ -293,7 +296,7 @@
+           break;
+         case 'b3dm':
+           try {
+-            let loader = new THREE.GLTFLoader();
++            let loader = new GLTFLoader();
+             let b3dm = new B3DM(url);
+             let rotateX = new THREE.Matrix4().makeRotationAxis(new THREE.Vector3(1, 0, 0), Math.PI / 2);
+             this.tileContent.applyMatrix4(rotateX); // convert from GLTF Y-up to Z-up
