+++ src/.eslintrc
@@ -116,5 +116,6 @@
       2,
       "all"
     ]
-  }
+  },
+  "parser": "babel-eslint"
 }
