+++ src/DESCRIPTION.ja.md
@@ -1,6 +1,8 @@
 このウィジェットは、MapLibre GL を使用したマップ・ビューア・ウィジェットです。
 
-## リファレンス
+## ソースコード
+
+ソースコードは https://github.com/lets-fiware/maplibre-gl-widget/ から入手できます。
 
 ## サードパーティ・ライブラリ
 
@@ -9,5 +11,9 @@ Maplibre GL ウィジェットは、次のライブラリを利用します:
 - [maplibre/maplibre-gl-js](https://github.com/maplibre/maplibre-gl-js) BSD-3-Clause License
 - [Geodan/mapbox-3dtiles](https://github.com/Geodan/mapbox-3dtiles) BSD-3-Clause License
 - [mrdoob/three.js](https://github.com/mrdoob/three.js/) MIT License
+- [visgl/deck.gl](https://github.com/visgl/deck.gl) MIT License
+- [visgl/loaders.gl](https://github.com/visgl/loaders.gl) MIT License
+- [uber-web/math.gl](https://github.com/uber-web/math.gl) MIT License
+- [Kanahiro/deckgl-gsi-terrain-layer](https://github.com/Kanahiro/deckgl-gsi-terrain-layer) MIT License
 
 依存ライブラリの依存関係はリストから省略しています。
