+++ src/DESCRIPTION.md
@@ -1,6 +1,8 @@
 This widget is a map viewer widget using MapLibre GL.
 
-## References
+## Source code
+
+You can get the source code from https://github.com/lets-fiware/maplibre-gl-widget/.
 
 ## Third party libraries
 
@@ -9,5 +11,9 @@ The Maplibre GL widget makes use of the following libraries:
 - [maplibre/maplibre-gl-js](https://github.com/maplibre/maplibre-gl-js) BSD-3-Clause License
 - [Geodan/mapbox-3dtiles](https://github.com/Geodan/mapbox-3dtiles) BSD-3-Clause License
 - [mrdoob/three.js](https://github.com/mrdoob/three.js/) MIT License
+- [visgl/deck.gl](https://github.com/visgl/deck.gl) MIT License
+- [visgl/loaders.gl](https://github.com/visgl/loaders.gl) MIT License
+- [uber-web/math.gl](https://github.com/uber-web/math.gl) MIT License
+- [Kanahiro/deckgl-gsi-terrain-layer](https://github.com/Kanahiro/deckgl-gsi-terrain-layer) MIT License
 
 The dependencies of dependencies have been omitted from the list.
