+++ src/index.html
@@ -3,9 +3,6 @@
     <head>
         <meta charset="utf-8" />
         <link rel="stylesheet" type="text/css" href="css/styles.css" />
-        <script src="lib/js/three.min.js"></script>
-        <script src="lib/js/GLTFLoader.js"></script>
-        <script src="lib/js/Mapbox3DTiles.js"></script>
         <title>MapLibre GL widget</title>
     </head>
     <body>
